refactor coverage script for github actions

+++ tests/coverage-finder.sh
@@ -3,31 +3,53 @@
 # description
 # Run all the operation to find code coverag and generate an html report
 
-#set -x #echo all commands
-make clean
-make
+# Get options
+while getopts "hb" opt; do
+    case $opt in
+        h)  # Choose whether or not to generate html results
+            HTML=true
+            ;;
+        b)  # Allow building and running
+            BUILD=true
+            ;;
+        \?)
+            echo "Invalid option: -${OPTARG}" >&2
+            exit 1
+            ;;
+        :)
+            echo "Option -${OPTARG} requires an argument." >&2
+            exit 1
+            ;;
+    esac
+done
 
+if [[ $BUILD == true ]]; then
+    #set -x #echo all commands
+    make clean
+    make
 
-# make euler1D-VL and mhd1D files
-repo_root=$(git rev-parse --show-toplevel)
-echo $repo_root
-cd ${repo_root}/euler1D-VL
-rm -rf *.o *.gcno
-g++ -std=c++17 -Wall -Wextra -Wpedantic -fasynchronous-unwind-tables -fexceptions -D_GLIBCXX_ASSERTIONS -g --coverage -lgtest -lgtest_main -I/usr/local/Cellar/googletest/1.11.0/include -L/usr/local/Cellar/googletest/1.11.0/lib -lpthread -c *.cpp
+    # make euler1D-VL and mhd1D files
+    repo_root=$(git rev-parse --show-toplevel)
+    echo $repo_root
+    cd ${repo_root}/euler1D-VL
+    rm -rf *.o *.gcno
+    g++ -std=c++17 -Wall -Wextra -Wpedantic -fasynchronous-unwind-tables -fexceptions -D_GLIBCXX_ASSERTIONS -g --coverage -lgtest -lgtest_main -I/usr/local/Cellar/googletest/1.11.0/include -L/usr/local/Cellar/googletest/1.11.0/lib -lpthread -c *.cpp
+    make
 
-cd ${repo_root}/mhd1D
-rm -rf *.o *.gcno
-g++ -std=c++17 -Wall -Wextra -Wpedantic -fasynchronous-unwind-tables -fexceptions -D_GLIBCXX_ASSERTIONS -g --coverage -lgtest -lgtest_main -I/usr/local/Cellar/googletest/1.11.0/include -L/usr/local/Cellar/googletest/1.11.0/lib -lpthread -c *.cpp
+    cd ${repo_root}/mhd1D
+    rm -rf *.o *.gcno
+    g++ -std=c++17 -Wall -Wextra -Wpedantic -fasynchronous-unwind-tables -fexceptions -D_GLIBCXX_ASSERTIONS -g --coverage -lgtest -lgtest_main -I/usr/local/Cellar/googletest/1.11.0/include -L/usr/local/Cellar/googletest/1.11.0/lib -lpthread -c *.cpp
 
-cd ${repo_root}/tests
+    cd ${repo_root}/tests
+
+    ./Tests.exe
+fi
 
 # generate initial report
 capture_directories=(--directory ${repo_root}/tests
                      --directory ${repo_root}/euler1D-VL
                      --directory ${repo_root}/mhd1D)
 lcov --capture --initial ${capture_directories[@]} --output-file coverage_base.info
-
-./Tests.exe
 
 # Generate test results
 # --capture = get the data
@@ -47,7 +69,7 @@ lcov --remove coverage_all.info "${exclude_patterns[@]}" --output-file coverage_
 
 lcov --list coverage_all.info
 
-if [[ $1 == "html" ]]; then
+if [[ $HTML == true ]]; then
     echo -e "\n\n===== Generate HTML ====================================================="
     genhtml coverage_all.info --output-directory Code-coverage-html
     open Code-coverage-html/index.html
